Refactor: Split MyRewardTransactions.tsx into components
Refactor MyRewardTransactions.tsx into smaller, reusable components to improve code organization and readability.
